backend/requirements.txt
asgiref==3.4.1
cffi==1.14.6
Django==3.2.6
django-cors-headers==3.8.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
zope.event==4.5.0
zope.interface==5.4.0
